Removed/renamed some more unused stuff
