+++ concurrent/shared/src/main/scala/cats/effect/concurrent/Deferred.scala
@@ -20,7 +20,6 @@ package concurrent
 
 import cats.effect.concurrent.Deferred.TransformedDeferred
 import cats.effect.kernel.{Async, Sync}
-import cats.effect.concurrent.LinkedMap
 
 import java.util.concurrent.atomic.AtomicReference
